Calculates the equilibrium saturation concentration of oxygen in water at the supplied conditions

Calculates the equilibrium saturation concentration of oxygen in water at the supplied conditions

calc_DO_sat(
  temp.water,
  pressure.air,
  salinity.water = u(0, "PSU"),
  model = "garcia-benson",
  ...
)

Arguments

temp.water
a numeric vector of water temperature in degrees Celsius, or a unitted object of water temperatures.

pressure.air
barometric pressure in millibars, or a unitted object of barometric pressure.

salinity.water
a numeric vector of salinity in PSU, or a unitted object of salinity. Defaults to zero.

model
character. One of 'garcia-benson', 'garcia', 'weiss', or 'benson', but 'garcia-benson' is recommended.

...
additional parameters passed to LakeMetabolizer::o2.at.sat.base

Value

a numeric vector of dissolved oxygen equilibrium saturation concentrations, in mg/L, with units attached if any of the input vectors are unitted.

Examples

calc_DO_sat(temp=21, press=1000.1, sal=0) # no units checking if no units provided

#> [1] 8.795956

library(unitted)
calc_DO_sat(temp=u(21,"degC"), press=u(1000.1,"mb"), sal=u(0,"PSU")) # units are checked

#> unitted numeric (mgO2 L^-1)
#> [1] 8.795956