streamMetabolizer: Models for Estimating Aquatic Photosynthesis and Respiration

streamMetabolizer uses inverse modeling to estimate aquatic photosynthesis and respiration (collectively, metabolism) from time series data on dissolved oxygen, water temperature, depth, and light. The package assists with data preparation, handles data gaps during modeling, and provides tabular and graphical reports of model outputs. Several time-honored methods are implemented along with many promising new variants that produce more accurate and precise metabolism estimates. This package is a fork of the original USGS-R streamMetabolizer package. It includes enhancements and bug fixes beyond the original version. The original package can be found at https://github.com/DOI-USGS/streamMetabolizer.

This package uses inverse modeling to estimate aquatic photosynthesis and respiration (collectively, metabolism) from time series data on dissolved oxygen, water temperature, depth, and light. The package assists with data preparation, handles data gaps during modeling, and provides tabular and graphical reports of model outputs. Several time-honored methods are implemented along with many promising new variants that produce more accurate and precise metabolism estimates.

Details

See https://usgs-r.github.io/streamMetabolizer for package vignettes.

Calculate new input variables

calc_depth() estimates stream depth from discharge.

calc_DO_sat() calculates dissolved oxygen saturation.

calc_light() models photosynthetically active radiation.

Convert existing input variables

convert_localtime_to_UTC() converts local time to UTC.

convert_UTC_to_solartime() converts UTC to local solar time.

convert_k600_to_kGAS() converts K600 to another gas.

convert_PAR_to_SW() converts photosynthetically active radiation to shortwave radiation.

Model metabolism

mm_name() chooses a model structure.

specs() sets model specifications.

metab() fits the model.

Inspect model results

predict_metab() predicts daily metabolism.

predict_DO() predicts dissolved oxygen.

plot_metab_preds() plots daily metabolism predictions.

plot_DO_preds() plots dissolved oxygen predictions.

get_params() extracts fitted and fixed parameters.

get_fit() extracts the internal fitted model.

get_mcmc() extracts Bayesian MCMC model objects.

get_fitting_time() extracts the model fitting time.

Inspect model inputs/properties

get_specs() extracts model specifications.

get_data() extracts subdaily fitting data.

get_data_daily() extracts daily fitting data.

get_info() extracts user-supplied metadata.

get_version() extracts the package version used to fit the model.

See also

Author

Maintainer: Connor Brown ConnorBrown1996@gmail.com (ORCID) [contributor]

Authors:

Alison P. Appling aappling@usgs.gov (ORCID)

Robert O. Hall (ORCID)

Maite Arroita (ORCID)

Charles B. Yackulic (ORCID)