Deprecated: Calculate a vector of dissolved oxygen deficits

Deprecated: Submit a GitHub issue if you want calc_DO_deficit() to stick around. Creates a DO.deficit vector for input into metabolism models.

calc_DO_deficit(DO.obs, temp.water, pressure.air, salinity.water = 0, ...)

Arguments

DO.obs
a numeric vector of dissolved oxygen concentration observations, mgO2 L^-1, or a unitted object of dissolved oxygen concentrations.

temp.water
a numeric vector of water temperature in degrees Celsius, or a unitted object of water temperatures.

pressure.air
barometric pressure in millibars, or a unitted object of barometric pressure.

salinity.water
a numeric vector of salinity in PSU, or a unitted object of salinity. Defaults to zero. Length must be one or equal to length of temp.water.

...
additional parameters passed to LakeMetabolizer::o2.at.sat.base

Value

a vector of DO.deficit values

Examples

if (FALSE) {
# Warning: this function is deprecated.
calc_DO_deficit(DO.obs=7, temp.water=25, pressure.air=900, salinity.water=2.43)
library(unitted)
calc_DO_deficit(
  DO.obs = u(c(7,7.5,7),'mgO2 L^-1'),
  temp.water = u(c(25,24.5,18.9), 'degC'),
  pressure.air = u(c(900,903,910), 'mb'),
  salinity.water = u(2.43, 'PSU'))
}