Extract any MCMC model objects that were stored with the model

A function specific to metab_bayes models. Returns an MCMC object of class stanfit (rstan::stanfit), which is saved in the metab_model by default because you should almost always inspect it; see keep_mcmcs argument to specs() for options for saving space. The rstan methods for rstan::stanfit objects include summary(), get_stancode(), stan_dens(), stan_diag(), and many more. See ?'rstan-plotting-functions', rstan::stanfit and the rstan manual.

get_mcmc(metab_model)

# S3 method for metab_bayes
get_mcmc(metab_model)

Arguments

metab_model
A Bayesian metabolism model (metab_bayes) from which to return the MCMC model objects

Value

The MCMC model objects

Methods (by class)

metab_bayes: Get the Bayesian MCMC model object