Describe the requirements for an argument to metab()

Describe the requirements for an argument to metab()

metab_inputs(
  type = c("bayes", "mle", "night", "Kmodel", "sim"),
  input = c("specs", "data", "data_daily", "info")
)

Arguments

type
the type of model you want to fit

input
the name of an argument to pass into metab()

Examples

metab_inputs('night','specs')

#> [1] "specs(mm_name('night')) # see ?mm_name, ?mm_specs for more options"

metab_inputs('bayes','data')

#>      colname          class          units     need
#> 1 solar.time POSIXct,POSIXt                required
#> 2     DO.obs        numeric      mgO2 L^-1 required
#> 3     DO.sat        numeric      mgO2 L^-1 required
#> 4      depth        numeric              m required
#> 5 temp.water        numeric           degC required
#> 6      light        numeric umol m^-2 s^-1 required
#> 7  discharge        numeric       m^3 s^-1 optional

metab_inputs('Kmodel','data_daily')

#>            colname   class    units     need
#> 1             date    Date          required
#> 2       K600.daily numeric     d^-1 required
#> 3 K600.daily.lower numeric     d^-1 optional
#> 4 K600.daily.upper numeric     d^-1 optional
#> 5  discharge.daily numeric m^3 s^-1 optional
#> 6   velocity.daily numeric   m s^-1 optional

metab_inputs('mle','info')

#> [1] "info may be NULL, a list, or any other data you want to attach to the output of metab()"